Migrate build system to Vite
This required re-implementing react-dropzone-uploader because it fails to build with the new system and the project seems abandoned.

This also adds ESLint to lint the project, as well as cleaning up all of the dependencies.

Speaking of dependencies, we can now use PostCSS 8, which means we can also use the latest TailwindCSS.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>